Moved execution environment headers to new file structure.
It was decided that the previous file structure was confusing because
it did not match the namespaces at all.  We are changing the directories
containing header files such that they always match the namespaces.

Another change we are making is that headers previously in "core" are
now in internal namespaces (and directories) underneath the interface
directories.  This subnamespace is considered part of the package of
the containing namespaces (so that the objects in each can point to
each other), but, as the name implies, should not be used by end users.