Move testing to its own directory
All the real testing (for now) is in CUDA directly.  I am keeping these
tests away from the rest of the CUDA code because putting it in the
same namespace does not seem like the right thing to do.

Later I plan to add unit tests, which test small internal components
and generally have to be "embeded" in a package to access internal
components.