Remove DataArrayStructuredConnectivity
Connectivity information now handled by the class objects.

This is a work in progress that does not fully compile.  The CUDA
testing code is not updated to not pass this data.  For that matter, I
need to add the mechanism to the work to return a cell (rather than
create a cell).