Template arrays, fields, field map work.  Fix derivative, interpolate.
Lots of restructuring here.  Had to move quite a bit of methods between
classes to get the dependencies to be noncircular.

This code has never been run, so it is likely to be not working.  I plan
to make unit tests for all the pieces.  Also, the cuda integration tests
are broken.  They are not changed for all the structural changes.  That's
also todo.