Do not keep references as object ivars.
Several classes (Cell, Field, WorkMapField) were internally holding a
reference to another object (that is, declared a member with &).  The
intention was to pass references rather than copy entire objects.
However, on retrospect this is a very unsafe practice.  If the object
being referenced goes out of scope, really weird and bad things will
happen.