Enable intermediate DataArrays
I added a field to DataArray that creates an offset for the index.  The
purpose of this offset is to enable temporary intermediate arrays.
Rather than allocate a large array that all threads share for values
that are just thrown away anyway, instead each thread creates a local
variable and creates a very small, offset DataArray using the pointer to
that variable.  As long as we are careful to give worklets only work
objects that can access that small part of the array, it should be fine
and relatively efficient.