Moved over to using lazy execution instead of immediate execution.
Execution of filters now happen when a filter is passed to a sink function.
The sink function will call compute on the last filter which will recursively
call compute on its dependent filters. Also added in a check that makes sure
modules that have been computed aren't recomputed when doing a fan out.