Adding support to track where arrays are located (control vs execution)
We don't want to move arrays from the execution enviornment to the host
after each worklet. Instead we want to keep them in the execution
enviornment intill somebody wants them in the control enviornment.

This is a step towards this feature. You automatically upload
arrays to the execution env when you ask for the execution array.
It won't make sure the modification to the execution array are
propagated down to the host if the host array already exists.

Now way to say delete execution array if you don't need it.