Refactored ArrayContainer, standarized the retrieve function.
Retrieve function pulls the ArrayContainer contents to the correct
area, be it control or execution. The user uses retrieve with
a dax::cont::Array to get the result from execution if it is there
or from the host if that is where it exists.

We still need a way to track when the user modifies the returned array
as we don't have proper const returns