Move VectorOperations to cont and exec namespaces.
The helper functions in VectorOperations was originally placed in the
dax namespace because they were equally applicable in both the execution
and control environment.  As with anything in the dax namespace, these
functions were given both __host__ and __device__ modifiers when
compiled with CUDA so that they could be called either from a worklet
(in a kernel) or from the control (in a CPU).

The problem is that these functions accept callback function/functors.
To avoid a CUDA compile error, the callback needs also to be given both
__host__ and __device__ modifiers.  That's no good since the callback
may only be needed in either the control or the execution.  To get
around the problem, these functions are replicated in the cont and exec
namespace so that they can be properly used in either environment.  To
avoid confusion I intentionally left the vector operations out of the
dax super namespace, but it could easily be added later if deemed
necessary.

The way I added these is a little wierd.  I wanted to write the code
once and share it in both namespaces to ensure the code was the same.
Normally I would just use preprocessor magic to do that, but I also
wanted doxygen to eventually document the functions correctly.  I
figured that would be easier if I used CMake's configure mechanism to
make two configured copies in each namespace, and then just have doxygen
work on the configured files.