Allow const functors in vector operations.
A recent previous change allowed you to pass in a functor reference so
that the functor object could hold and return state.  However, this
meant you could not pass in a const functor, which was useful when
creating the functor in the command line.

This commit adds versions of the functions that accept const versions of
the functor.  The only downside of this solution is that to use them you
have to declare the operator() as const.  This is really the right thing
to enforce anyway, since any modifications to the functor will get
confusingly lost.  However, there will be confusion with the error,
which will simply say that no instance of the function matches.