Allow CudaParameters to be declared in multiple object files.
Previously, CudaParameters declared a couple static members in its
header file.  This meant that any object file that included the header
declared a copy of these static members, which in turn caused linker
errors.

This fix moves the static members as static variables in a method.
There they become anonymous and can be replicated without linker errors.
The downside is that the queries will be repeated if called from a
different object file, but that is not a big deal.  The only other way
around the problem is to create a library containing the variables.