Determine CUDA threads/blocks independent of point/cells.
Previously, the CudaParameters class took a grid and recorded the blocks
and threads to use if mapping either the points or cells.  This turned
out to be more trouble than it helpped as several mappers can work on
either points or cells.  Now, it just excepts a number of items, and the
calling function must determine how many that is.