Worklet kernels must copy work object for OpenMP.
When running in OpenMP, the worklet kernels all share the same
parameters in memory.  They were previously all setting the index in the
same work object and often stomping over each other.  This change makes
them all do a local copy of the work object and set the index from
there.

This type of error is exactly why we want Dax in the first place.  It
took me 3 hours to find, and I was lucky to find it that fast.

Passing a work in the worklet parameters object may not be the best
idea.  It might be better to construct the work in the kernel.  That
would force us to have a seperate copy in every thread.  I want to wait
for other cell types before committing to that.