Changing scheduling to include error handling.
Schedule function now supposed to pass an ErrorHandler class to its
functor and then return an error string if an error was raised anywhere.

This is very much a work in progress.  The new scheme is only
implemented in the debug scheduler.  The work classes and worklet
launcher functions are not yet updated to use the ErrorHandler.