Added control version of test helper.
Only real difference is that it has a special catch for Dax exceptions.
Updated all tests in cont directory as well.

Also added a test_equal (in the global namespace) to
dax/internal/Testing.h.  This tests the equality of scalars and vectors
allowing for some variance from expected floating point errors.