Schedule functions throw exception rather than return string.
Previously, the schedule function in DeviceAdapters would return a
string if an error was raised in the execution environment.  It was up
to the calling program (generally the worklet scheduling function) to
check the string and raise an error.  I thought this would make creating
schedule functions easier, but it hasn't and it complicates everything
else.  Instead, have the schedule function throw the error itself.