Use "safer" copies of error string in thrust schedule.
Rather than use thrust iterators in the thrust schedule code to identify
and get errors, use the already existing CopyTo/From methods in
ArrayContainerExecutionThrust.  The effect should be the same, but it
minimizes the interface points to thrust.

This is a (perhaps lame) attempt to track down a problem we are having
with some of the Cuda tests that happen to involve the ErrorExecution.
It's a long shot, but perhaps Thrust did not like how I was using its
iterators.