Clear Cuda error when throwing out of memory.
When the ArrayContainerExecutionThrust (Cuda or otherwise) failed to
allocate an array, it properly caught thrust's exception and threw its
own.  However, the Cuda version of Thrust did not clear out the Cuda
error before throwing its error.  This change catches the out of memory
exception, checks the Cuda error, and clears it if it is the same out of
memory error that apparently caused the currently thrown exception.