Add the new code to generate a stencil on the device adapter.
By extracting the point mask work from the workremovecell it was expected
that the performance would improve. The actual result is that the performance
has decreased. The reasoning seems to be that the bottleneck in the the worklet
isn't actually the non-coalesced memory access but something else in the worklet.
I am thinking the actual issue is the number of registers being used.

The timing differences is the the worklet is still taking 0.2 ms to run,
but now we are adding on 0.08 to that for executing the point mask creation.
Even worse is that the new point mask creation algorithm doesn't scale as well
as the old naive aproach when we have a higher ratio of cells being kept around.