Identify year in copyright statement and change to current year.
This makes it easier for people using this test in several ways.  First,
there are no regular expressions given in either DaxCopyrightStatement.txt
nor in the suggested insertion.  Second, the correct date is automatically
inserted, so you can just cut and paste.  Third, if we want we could add
any regex special characters and escape them and still have the script
work.