Require setting Dax_SOURCE_DIR when calling DaxCheckCopyright.cmake.
I learned (probably for at least the second time) that
CMAKE_CURRENT_SOURCE_DIR does not do what you expect when executing
a CMake script with the -P command.  I can find no way to get the
filename or path of the current script, so instead force setting it
as an input variable.