Simplified the threshold worklet, and made it more inline with the paper.
Instead of using hand unrolled code we moved to calling
vector for-each which is smarter, since that can be later made faster
in one place. Also this API is what we talk about in the threshold paper
as it removes the onus from the developer to get correct loop iteration.