Redesign Field classes for const correctness.
The previous version of the Field classes did not perserve const
corrected.  It was technically possible for a worklet to write to an
input field, which is bad.

This version of the Field classes has duplicate versions for input and
output.  Internally this is handled with a base field class that is
templated in a quasi-policy-based design.  An access policy determines
what read/writes are allowed and what type of iterator is held
(const/non-const).  Because the syntax of defining this base field is a
bit onerous, there are trivial subclasses for all the expected
combinations.  Although these classes no longer inherit from each other,
internal classes can apply to particular subsets of fields with clever
templating and definitions of the base field.