Move topology structures to dax::exec::internal namespace.
As we remove the dependence of DataArray, some of these structures will
start to depend on features in the exec environment.  Although it is
convienient to have one structure shared between the two environments,
as soon as we need to access arrays (which is the case for all but the
uniform mesh), you can't really share anyway.