Make work objects const.
Previously, work objects were mutable in that you could, for example,
change the index with which they were working.  This opened the
possibility for a worklet to do something dangerous like changing the
index to access data for another cell.

Things like current indices are now set at creation and fixed.  Also,
the topology is hidden so that you can't query that to jump to some
inappropriate part of memory.