Remove TopologyUnstructured's dependence on ExecutionAdapter.
One major ramification of this is that CellHexahedron can no longer
directly depend on TopologyUnstructured because that would require the
cell class to be templated on ExecutionAdapter, which would be a major
pain.  This means that CellHexahedron cannot read arrays to find
connections.  Instead, on creation CellHexahedron copies all of the
point connections to an internal tuple.  Since the API was recently
changed to return all point connections as a tuple anyway, I suspect
this will not impact performance much.

In a related note, the cell class has also become more constant.  Once
the fields are set in the constructor, they are not allowed to change.
Thus there is nothing like SetIndex, which was dangerous anyway.