Make error handling dependent on ExecutionAdapter.
The old method always wrote the error message to a buffer and then
checked that.  This method puts a class in the ExecutionAdapter that has
a method to raise the error that can either do that or throw exception.
In addition to better taking advantage of available device features, the
previous version required allocating an array that possibly used the
user's semantics, which could have been difficult.