Update UniformGrid and UnstructuredGrid to new topology classes.
Also cleaned up the interface by deleting several methods pertaining to
getting data in the control environment.  Since ArrayHandles are not
really set up to get data in the control environment, this was a hack to
implement.  Furthermore, there is not much purpose for it.  It is mostly
to implement testing.