Change unstructured grid topology to not hold point coordinates.
This was redundant information as you needed a point coordinate field
object to get the point coordinates, and that field object already has
the array of point coordinates.  It was also wasteful when running a
worklet that did not need point coordinates.

Also update getting point coordinates in the WorkMapCell class, which
depends on these changes.