Fix an issue with LowerBound and in place arrays.
The in place version of DeviceAdapterSerial::LowerBound was not properly
calling PrepareInPlace on the values/output array.  Instead it was
calling PrepareInput and PrepareOutput, which is wrong.

On a similar note, I discovered a problem with PrepareInPlace: it does
not work if the ArrayHandle has user-supplied iterators.  This is
because there is no array allocated on the control, and even if there
was it would have to be copied in a shared environment.  In the end, I
just raised an exception to specifically disallow this.