Check assert in conditions in macros.
Previously, the macro just called a function that checked the condition.
Even when this was inlined, this caused issues in the control environment
because the function accepted std::string arguments, which caused those
string arguments to be created from c strings even when the condition
passed (as it always does).

To fix this, the assert macros now check the condition before calling
the function.