Remove overloading from templated ExecutionPackageField.
The ExecutionPackageField function had some overloading placed on top
of some fairly complicated templating.  This was working for GCC 4.2
but not so much for GCC 4.6.  Perhaps function or template matching
rules relaxed to allow more matching and that opened up ambiguity.

At any rate, I solved the problem by removing the overloading that was
not really template specialization.  The ExecutionPackageField functions
that deal with a given array size or are automatically sized by a grid
are differentiated with the suffix Array or Grid.  Furthermore, those
functions that accept a const version of ArrayHandle, which are assumed
to only allow inputs, have a Const suffix.  I'm not super happy with this
solution, especially with respect to the const differentiation, but it
works.