Fix "cc" clobber and unused variable warnings with boost/nvcc
I traced down these problems to some specialized code in the boost smart
pointer that used special tricks of the x86 architecture (or other
instruction sets as appropriate) to atomically reference count with
threads.  The problem was that the nvcc compiler was not processing the
assembly block the same way as gcc (even though boost somehow identified
nvcc as gcc).  That was what was causing the errors.  I'm not sure if
the code was actually being generated correctly.

The easiest way around this problem is to specify how the shared pointer
should handle threads by defining one of three macros:
BOOST_SP_DISABLE_THREADS, BOOST_SP_USE_SPINLOCK, BOOST_SP_USE_PTHREADS.
Unfortunately, I am uncomfortable with specifying any of these in the
Dax header files for fear that some user will legitimately expect the
behavior to work in a different way then I choose.  So my solution is to
check to see if one of these is set when compiling under CUDA and issue
my own verbose warning if not.  Granted it is one more annoying (perhaps
more annoying) error, but it is very descriptive and tells how to get
rid of it and the other errors.  I've also defined
BOOST_SP_DISABLE_THREADS in the header test builds and manually added
that definition to the CUDA unit tests and testing code.