Get rid of double usage in test_equal.
This gets rid of an nvcc warning about using doubles.  This originates
from compiling for the 1.0 compute capabilities.  You might say who
cares about 1.0 compute capabilites? but it is still a nice check to
make sure we are not using doubles when floats are set for the scalar
type (since double computes take longer, particularly on GPUs).  Thus,
let's clean these errors out of all the code.