Update traits to work on generic Tuple objects.
Also update basic arithmetic operations (such as +,-,*,/ operations and
dot products) to work on tuples.  I left the specializations in place
because they might be faster (they all have unrolled loops), but we
might want to change the specializationto work on Tuple types instead if
they get used a lot.