Fix issue with finding parametric coordinates in hexahedra.
The problem was that the Jacobian returned from the function in
derivative function was actually returning the transpose of the
Jacobian.  This has been fixed and the derivative function is now well
documented to declare why it is using the parametric coordinates and how
it works in the first place.