Added daxNotUsed to specify unused parameters.
There are several instances in Dax in which a function or method
parameter is intentionally unused.  We have a compiler warning that will
alert you if a named parameter is unused, and it helps identify errors.
This macro works the same way as vtkNotUsed.  It is an empty macro that
you can surround the parameter name with.  It allows you to name what
the parameter is for without getting grief from the compiler.  As an
added bonus, you give a human indication that the parameter is
intentionally not used.