An experiment of writing the square worklet without any field classes.
Instead the field values themselves are passed directly to the worklet
function.  After some discussion during Dax design meetings, this has
the potential for simplifying both the worklet function itself and the
invocation method.  It also has the potential for expanding the
capabilities of Dax (for example, supporting multiple different types of
array accessors).