Added ArrayPortal concept and ArrayPortalFromIterators class.
The ArrayPortal concept is like an iterator except that it has an
interface like a container.  This makes it much easier to implement new
containers than new iterators.

I have also added an ArrayPortalFromIterators class.  This class
provides a concrete templated ArrayPortal class that access the same
data as a specified set of begin/end iterators.