Fix dependency issue with thrust DeviceAdapter.
The CUDA/OpenMP tag was not fully declared before ArrayHandle was
loaded.  This caused problems when compiling a (templated) function that
then tried to use the tag (at least with the CUDA compiler).

This is just a temporary duck tape fix.  There are generally more
serious issues with the dependencies.  A more complete fix is proposed
on the wiki:

   http://www.daxtoolkit.org/index.php/DeviceAdapter_Dependency_Issues

However, this fix is too extensive to perform on this topic branch but
would cause lots of bad conflicts if done on a separate branch.  Thus,
the quick fix now and a more extensive fix later.