Make UniformGrid return an implicit ArrayHandle for points.
This gets around a lot of problems with treating the points for uniform
grids specially.  However, this can only work if a worklet can be
applied to ArrayHandles with different containers.  Currently it cannot,
so a lot of code is broken.  I plan to look at and convert worklets that
use coordinates next anyway.