Convert elevation to use no field objects.
This demonstrates the passing of the implicit array of coordinates of a
uniform grid to the elevation filter.  As such, it also demonstrates a
worklet accepting/generating values for arrays of different containers
without having nasty templating.  (In fact, there is much less
templating.)

It should be noted that the implicit array is imperfect in that if this
were to be run on, say, a GPU, an array with the values for the implicit
array would be created on the device.  We'll need some template magic to
fix this later.