Nuke Field objects from orbit.
With the experimental changes of using worklets without passing
field objects, the field objects become basically unused.  This
change nukes the field objects from orbit.  There were (and are)
still a lot of dependencies to them, so those get removed, too.

The current state is that the Field object and all its depedencies
in the execution environment are removed and all the execution
environment unit tests pass.