Blow away execution adapter
The main function of the execution adapter was to hold the type of array
portal (previously iterator) used for fields.  Field objects no longer
exist and worklets do not need to know about the array portal.  This is
a double win in that the worklets do not need to be templated for this
and that each field can have a different type of array portal.

The only other thing the execution adapter provided (and the only thing
I can think of it usefully providing right now) is a customized error
handling routine.  However, except for in tests and serial scheduling,
they all just used an array buffer.  I don't think this functionality is
worth propogating the template parameter everywhere for, so bye-bye.