Destroy the work objects.
Since we are working on scorched earth policy, the work objects are now
destroyed.  They are replaced with worklet objects.  As suggested by (I
think) Brad King, a worklet implementation can subclass one of these
objects and implement an operator() to make it a functor.  This serves
the double purpose of having all the "work" facilities incorporated (the
only real one so far is the RaiseError method) and to identify worklet
type by template programming.  There is also the fringe benefit of
having state associated with the worklet object to make it easier to
establish worklet parameters (such as isosurface value).