DeviceAdapter algorithms to handle multiple containers.
All the DeviceAdapter algorithms were designed to have all the arrays to
have the same container types.  That is no longer a constraint, so the
templates have been broadened to accommodate this possibility.

The thrust DeviceAdapter algorithms still need to be fixed.  I have not
gotten that far in the build.