Return reference to arrays in UnstructuredGrid.
The UnstructuredGrid class has ArrayHandles for the cell connections and
point coordinates, respectively.  This change makes the Get methods
return a reference to these arrays (in both const and non-const
versions) so that they can then be passed to functions that take
references to ArrayHandle objects.

This may not be the best solution.  On retrospect, this is an
ineffective attempt to add const-correctness to ArrayHandle.  A more
appropriate method is to have a const version of ArrayHandle.