Add special path for gradient worklet on uniform grid.
You do not need the point coordinates to compute the gradient on voxels,
unlike other worklet types.  However, since the field objects have been
removed the point coordinates are loaded before the worklet is executed
whether they are used or not.  To get around this problem, I added a
specialized version of gradient for voxels that does not take point
coordinates as a parameter.