Make functors for the DeviceAdapter Schedule constant.
Previously, the functor kernels that executed worklets were non-const so
that they could modify their Worklet ivar to use the ErrorMessageBuffer
passed to it.  This is problematic because multiple threads are trying
to set the same entry, which can wreck havoc on the caching system.  It
is also conceptually a bad idea to allow the kernels to be modifying
their own ivars as it invites other read/write collisions.