Select ArrayContainerControl with macro.
Previously the array container was selected by including the appropriate
header file, which automatically set a default macro of its own type.
The problem with this is that it caused problems with the order of
includes, and it was impossible to include a secondary container type
without overriding the default and potentially getting weird behavior.

Also fixed here are some instances of using default values when they
shouldn't be (albiet only in test code).  This was already supposed to
be done, but it was not included right (another reason not to use the
old version).