Select DeviceAdapter with macro.
Previously the device adapter was selected by including the appropriate
header file, which automatically set a default macro of its own type.
The problem with this is that it caused problems with the order of
includes, and it was impossible to include a secondary container type
without overriding the default and potentially getting weird behavior.

This is a work in progress in that it does not compile.  A dependency
issue has been hit, which I will address next.