Break DeviceAdapter into pieces.
The device adapter is now broken into three pieces: a tag, an
ArrayManagerExecution class, and a set of algorithms.  This is so that
the ArrayHandle header can depend only on the tag and
ArrayManagerExecution and the algorithms can depend on the ArrayHandle
without creating a dependency loop.

The design is currently documented on our Wiki at
http://www.daxtoolkit.org/index.php/DeviceAdapter_Dependency_Issues

Only the serial device adapter has been updated.  The thrust based
adapters will follow shortly.