Move DAX_DEVICE_ADAPTER* macros
The definition of all the DAX_DEVICE_ADAPTER_* values and the default
setting of DAX_DEVICE_ADAPTER is moved from DeviceAdapter.h to
DeviceAdapterTag.h.  This is to facilitate when specific device adapters
include DeviceAdapterTag before DeviceAdapter is included.  Previously,
not all the appropriate macros were set in time and the wrong default
device was selected.