Handle exceptions in OpenMP execution environment.
Normally, exceptions are not allowed in the execution environment.
However, the OpenMP device adapter shares ArrayPortal objects between
control and execution environments.  Because these shared arrays are
designed for the control environment, they might throw an exception.
That is mostly OK because OpenMP runs on a CPU that can handle it.
However, exceptions have problems when propogated past a thread
schedule, so catch it and handle as normal.