Made comparison operators const.
The big concern is if the comparison should be done by loops or
with template recursing. I need to look at the assembly generation and
the performance on the gpu. I do know that for loops incur a slowdown on the
gpu, I just don't know if the template method unrolls properly at compile time.