Add install targets for all header files.
This makes it possible to do a make install that will put all header
files in one location that you can point to with an external program.

This work also adds some CMake files in the worklets directory so that
these get added to the installed header list (which is necessary to
compile code).  This further highlights the stupidity of the current
location of this directory.  As a side effect of these changes, the
Worklets show up in an IDE file list.