Add configure macros for external projects.
To simplify the use of Dax from external projects, include in the
install a set of cmake files that will configure the build for various
devices (Serial, OpenMP, CUDA).

Now when an external project calls find_package(Dax), this defines
macros named DaxConfigureSerial, DaxConfigureOpenMP, and
DaxConfigureCuda.  These will all find and set up the dependent packages
(such as Boost and Thrust).  They will set the macro Dax_<device>_FOUND
to true if the load was successful.  They all take the optional argument
REQUIRED, which will cause a CMake error if the dependent package could
not be loaded.