Fix conflicts between std and boost::math functions
For technical reasons explained in a comment, certain math functions use
the boost version.  These boost functions often have similar functions
in the std namespace from the compiler-provided functions.

The Dax header file used the using keyword to point to the boost math
functions (which probably was not a great idea in the first place).  If
somewhere else using namespace std was declared, then it would be
ambiguous which function to use.  The problem was resolved by moving the
using declaration inside the function it was needed.