Update the tests to understand the new unstructured grid template parms.
The unstructured grid now can have different containers for points
and topology so that needs to be propagated to the device adapter and
grid generator code.

I also used this to refactor a bit of the code in the device adapter
to use the grid generator. This way we only have a single place where
will list all the grid types to test a functor on.