Add dax::exec::arg::BindCellPoints for WorkletMapCell
WorkletMapCell schedules for cells so indexes refer to cells.  If a
Field argument in the ControlSignature is tagged as a Point field then
it cannot be directly indexed.  Instead we need to use a Topology
argument to map the cell index to point indexes and pass the field falue
from every point to the worklet in the execution environment.  Add
BindCellPoints to represent this operation.  Add a dax::exec::arg::Bind
specialization for WorkletMapCell to choose between BindDirect and
BindCellPoints based on the parameter tags.