Add dax::cont::arg::Field mapping for ArrayHandles.
We convert array handles to field portals. This requires that the
scheduling code will have to tell each array handle that has the
output tag what size to allocate.

In the future we have to also add an ability to all concept maps
to support bringing results back to the control env.