Add in the infastructure to schedule abitrary worklets.
This just creates the exec fields needed for the worklets to execute.
It doesn't actually execute the worklets. The outstanding issues
are that we need to find the current device adapter and
saving fields with the Out tag.

We also need to convert the header to use boost preprocessor macros
to handle abritrary number of arguments.