Members.h now has custom versions for control and exec.
When compiling with the cuda device adapter Members.h on the control
side will be binding to control only exported classes. This will cause
a confliction since it is exported as host and device. The same happens
in the execution enviornment, where we get warnings of calling a
device only call from a host/device method.

The quick solution is to configure a version of the members header
for both control and exec. While this works, it does create some
heavily complex code inside Member.h.in. In the future I want
to look at just having too versions.