Schedule: Provide reference implementation in C++11
Convert the single-argument C++03 reference implementation to C++11.
Rename "Schedule.txx" to "Schedule_Cxx03.h" to clarify that it provides
the C++03 implementation.  This approach leaves only one implementation
for each version of C++.

While at it, simplify the implementation to put the logic directly in
the constructor.  This avoids an extra layer of forwarding/copying the
arguments.