Remove the InOut tag as that is not behavior we want in dax.
The InOut tag might be really simple for a vector or an array handle,
but when it comes to topology and other more complicated types the
tag causes more problems that it is worth.

If you want to modify an existing arrayHandle, just pass it to a worklet
as the input and output parameter. It won't cause any duplicate
memory in the execution enviornment and you get the same results
as the InOut tag.