Convert GenerateTopology to use some of the new schedule infastructure.
We promote ScheduleGenerateTopology to be a dax::cont Class that
cam be used by anybody. Use pass in a classification array and
it will handle doing the topology generation.

Currently we are limited to the classify step only being passed in
the input and output grid, as it doesn't use the new infastructure.

The upside is that the new infastructure can threshold other arrays
once it is done.