Now able to Map a parameter using key/value lookup.
If you have a argument that you don't want to lookup on using the current
workId, you now can use ScheduleMapAdapter to use a custom lookup table.
The keys must be the same length as the number of elements you are dispatching on.

that the worklet will get the value at value[key[workId]].