Have to move schedule implementation to operator()
When the implementation of schedule is in the class constructor we get warnings
that we are not using the object. Really creating an object and doing the work
in the constructor was a bad design to start with.

The side benifit of this way is now you can construct a single instance
of dax::cont::Schedule with the desired device adapter and pass it around
calling it like a function to run a worklet, which looks really nice.