Attempt at choosing correct thrust iterator.
The problem is that now that the ArrayTransfer class can specialize the
type of portal on the execution side, the execution portal is not always
the expected portal from the ArrayManagerExecution.  Now you have to
choose what iterator to use based on the portal we get.

This is a first attempt, but it is not right as I get an ambiguous
resolution of templated/overloaded functions.