Choose correct thrust iterator.
It uses a thrust::device_ptr if the iterator can be resolved to a
pointer or to a thrust::transform_filter that reads data from an
ArrayPortal otherwise.

The code compiles (for OpenMP) and the tests pass (although I have not
checked the speed or to make sure the scheduling happens as expected).
Even if this works, I'm not particularlly happy with the solution.  The
problem is that the non-pointer iterators are read only.  This means
that any writes to an indirect pointer (such as for the VTK adapter)
will fail.