Fix issue where a shared array was not resized properly when shrunk.
While debugging another problem, I noticed that when using a DeviceAdapter
that uses ArrayManagerExecutionShareWithControl, there was a problem
with shrink.  When you called shrink on the ArrayManagerExecution, it
would just shrink its local array portal.  The control array that the
portal points to was never resized.

This change adds a shrink call to the control array when the output
data is retrieved.  This is a good time to do it since the ArrayHandle
will consider the control array invalid until this output retrieval is
called.