Add test for writing to funky iterators with thrust.
The recent changes for ArrayTransfer necessitated that if the iterator
was not a raw pointer, then the thrust device adapters created its own
iterator to force thrust into scheduling on the device.  This test makes
sure that this work properly (at least for correctness) for writing.

And the test currently fails because this writing is not supported.