Make custom thrust iterator writable.
Rather than return a value from the functor used with the thrust
transform iterator, it returns a reference object that can be used to
set or get the value.  This allows you to use this iterator as an output
iterator.

(I'm a little surprised the compiler lets me do this.  I would have
expected the transform iterator to be declared as input-only, but I
guess it is not.)