Make TopologyCell return by reference, and be able to save cells.
Before we redesigned cells to have an empty constructor it was
impossible to return a cell by reference, which caused massive slowdowns
in worklets that needed cells as a parameter. Now we can store
a cell reference TopologyCell becomes a far better class.
Now you can save a cells topology directly to a cell and not
have to use the glue wrapper PointIds.