Fix warning about use of uninitialized value.
This was hard to track down.  Deep in some templated code there is an
object that (potentially) holds a Tuple as an ivar.  The default
constructor for the Tuple is used, which does not actually set the
values in the internal array.  I'm pretty sure what happens next is that
this class is magically copied somewhere, and the compiler traced the
copy to the uninitialized value.