Creating general device adapter algorithms (starting with StreamCompact)
Currently there is a major compile problem.  The general algorithm needs
to call the correct algorithm for the device adapter.  Currently the
compiler is giving me grief about not finding the appropriate symbols
for the device adapter algorithms that are implemented.

It is not totally clear to me why I'm getting problems here and not
elsewhere.  I believe the issue is that the device adapter algorithms
are implemented as overloaded templated functions.  When compiling the
generic function, it grabs the implementation of the overload that it
sees not knowing about the other overloads.  Even though the correct
overload is defined when the templated function is instantiated, it was
not in the symbol table when the templated function was first compiled.

I can think of two possible ways around this problem.  First, I could
add the functions that need to be called as template parameters to the
generic functions.  Second, I could change all the device adapter
algorithms to be inside a specialization of a templated struct.  I only
expect to do the latter if this issue comes up in a way the user can
see.